CI-22: minor typo fixes in docs, installation from directories fixed, comments added

+++ lib/Netresearch/Source/Filesystem.php
@@ -4,11 +4,21 @@ namespace Netresearch\Source;
 use Netresearch\Source\SourceInterface;
 use Netresearch\Source\Base as Source;
 
+use \Exception as Exception;
+
+
 class Filesystem extends Source implements SourceInterface
 {
     public function copy($target, $branch = 'master')
     {
-//         $command = sprintf('rsync -a -h --exclude="var/*" --exclude="*.git" %s %s 2>&1', $this->source, $target);
+        if (!file_exists($this->source)) {
+            throw new Exception("Source directory does not exist: '{$this->source}'");
+        }
+        
+        // make sure that source ends with directory separator
+        // as we want to copy its contents, not the directory itself
+        $this->source = rtrim($this->source, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR;
+        
         $command = sprintf('rsync -a -h %s %s 2>&1', $this->source, $target);
         exec($command, $result, $return);
